Hot fix. Sorry for it :)

+++ app/src/main/java/ru/terrakok/gitlabclient/model/repository/todo/TodoRepository.kt
@@ -95,7 +95,7 @@ class TodoRepository @Inject constructor(
         )
     }
 
-    fun markPendingTodoAsDone(id: Int) = api.markPendingTodoAsDone(id)
+    fun markPendingTodoAsDone(id: Long) = api.markPendingTodoAsDone(id)
 
     fun markAllPendingTodosAsDone() = api.markAllPendingTodosAsDone()
 }
